Added new signals read_done and write_done
When the master NACKs after a read byte read_done goes high for 1 clock.
If the master sends START or STOP following a write byte, write_done goes
high for 1 clock.
Updated the test bench to add the new signals to the submodule, but the
new signals are not tested.  Looked fine in the waveform.
The Makefile points to my home directory ghdl path for now.